add remaining methods to compatibility layer
